# gpt-sovits-backend/GPT-SoVITS.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GPT-SoVITS",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/voice-clone/samples": {
      "get": {
        "summary": "获取用户的音频样本列表",
        "deprecated": false,
        "description": "获取用户的音频样本列表",
        "operationId": "get_voice_clone_get_user_samples",
        "tags": [
          "voice_clone"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "查询参数: page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "查询参数: per_page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/voice-clone/samples/{sample_id}": {
      "delete": {
        "summary": "删除音频样本",
        "deprecated": false,
        "description": "删除音频样本",
        "operationId": "delete_voice_clone_delete_sample",
        "tags": [
          "voice_clone"
        ],
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "description": "路径参数: sample_id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/voice-clone/start-training": {
      "post": {
        "summary": "开始训练",
        "deprecated": false,
        "description": "使用上传的音频样本开始语音克隆训练",
        "operationId": "post_voice_clone_start_training",
        "tags": [
          "voice_clone"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model_name": {
                    "type": "string",
                    "description": "模型名称"
                  },
                  "sample_ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "音频样本ID列表"
                  }
                },
                "required": [
                  "model_name",
                  "sample_ids"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "estimated_completion": {
                          "type": "string"
                        },
                        "sample_count": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        },
                        "task_id": {
                          "type": "string"
                        },
                        "total_duration": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "estimated_completion",
                        "sample_count",
                        "status",
                        "task_id",
                        "total_duration"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data",
                    "message",
                    "success"
                  ]
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/voice-clone/tasks": {
      "get": {
        "summary": "获取用户的音色克隆任务列表",
        "deprecated": false,
        "description": "获取用户的音色克隆任务列表",
        "operationId": "get_voice_clone_get_user_tasks",
        "tags": [
          "voice_clone"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "查询参数: page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "查询参数: per_page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "查询参数: status",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/voice-clone/tasks/{task_id}": {
      "get": {
        "summary": "获取任务详情",
        "deprecated": false,
        "description": "获取任务详情",
        "operationId": "get_voice_clone_get_task_detail",
        "tags": [
          "voice_clone"
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "路径参数: task_id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/voice-clone/tasks/{task_id}/cancel": {
      "post": {
        "summary": "取消训练任务",
        "deprecated": false,
        "description": "取消训练任务",
        "operationId": "post_voice_clone_cancel_task",
        "tags": [
          "voice_clone"
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "路径参数: task_id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "examples": [
              false
            ]
          },
          "message": {
            "type": "string"
          },
          "error_code": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}